Add SQLite backend and database module
