+++ qmtl/gateway/__init__.py
@@ -1,7 +1,8 @@
 from .dagmanager_client import DagManagerClient
 from .queue import RedisFIFOQueue
 from .worker import StrategyWorker
-from .api import create_app, Database, StrategySubmit, StrategyAck, StatusResponse
+from .api import create_app, StrategySubmit, StrategyAck, StatusResponse
+from .database import Database
 from .fsm import StrategyFSM
 from .ws import WebSocketHub
 from .watch import QueueWatchHub
