+++ tests/reliability/test_worker_alerts.py
@@ -5,7 +5,7 @@ from fakeredis.aioredis import FakeRedis
 from qmtl.gateway.worker import StrategyWorker
 from qmtl.gateway.queue import RedisFIFOQueue
 from qmtl.gateway.fsm import StrategyFSM
-from qmtl.gateway.api import Database
+from qmtl.gateway.database import Database
 
 
 class FakeDB(Database):
